Move Result classes/traits to Result namespace
See #22
