+++ src/ExpoPush.php
@@ -11,8 +11,8 @@ use Dru1x\ExpoPush\PushTicket\PushTicketCollection;
 use Dru1x\ExpoPush\Request\GetReceiptsRequest;
 use Dru1x\ExpoPush\Request\RequestExceptionHandler;
 use Dru1x\ExpoPush\Request\SendNotificationsRequest;
-use Dru1x\ExpoPush\Results\GetReceiptsResult;
-use Dru1x\ExpoPush\Results\SendNotificationsResult;
+use Dru1x\ExpoPush\Result\GetReceiptsResult;
+use Dru1x\ExpoPush\Result\SendNotificationsResult;
 use Generator;
 use Saloon\Exceptions\InvalidPoolItemException;
 use Saloon\Http\Pool;
